+++ src/main/java/br/com/webmail/enums/EnumFiltro.java
@@ -1,17 +1,17 @@
 package br.com.webmail.enums;
 
 public enum EnumFiltro {
-	CAIXA_ENTRADA("Caixa de entrada",1),LIXO("Lixo",2),RASCUNHOS("Rascunhos",3),EXCLUIDOS("Excluídos",4);
+	CAIXA_ENTRADA("Caixa de entrada",1l),LIXO("Lixo",2l),RASCUNHOS("Rascunhos",3l),EXCLUIDOS("Excluídos",4l);
 	
-	private Integer valor;
+	private Long valor;
 	private String descricao;
 	
-	private EnumFiltro(String descricao,Integer valor){
+	private EnumFiltro(String descricao,Long valor){
 		this.valor = valor;
 		this.descricao = descricao;
 	}
 	
-	public Integer getValor() {
+	public Long getValor() {
 		return valor;
 	}
